requirements.txt
#Flask>=0.11.1
Flask==0.11.1
Flask-Assets==0.12
cssmin==0.2.0
jsmin==2.2.1
flask-recaptcha>=0.4.1
flask-kvsession==0.6.1
flask-s3==0.3.1
flask-mail==0.9.0
flask-caching==1.0.0
flask-cloudy>=1.0.1
flask-seasurf==0.2.2
flask-babel==0.11.2
flask-cors==3.0.2
Active-Alchemy>=1.0.0
Paginator>=0.4.1
six>=1.9.0
passlib>=1.7.0,<1.8.0
bcrypt==3.1.3
python-slugify==1.2.1
humanize==0.5.1
redis>=2.10.0,<3.0.0
ses-mailer>=0.13.0
markdown==2.6.2
inflection==0.3.1
pyyaml==3.11
click==6.2
sh==1.11
dicttoxml==1.6.6
arrow==0.8.0
blinker==1.4
itsdangerous==0.24
requests>=2.13.0
